update nohostfs and env_posix.cc. but many errores occur until now

+++ nohost/nohost_fs.cc
@@ -2,19 +2,142 @@
 #include <string.h>
 
 namespace rocksdb{
-int NoHostFs::Close(int fd) {return 0;}
-int NoHostFs::Rename(std::string old_name, std::string name) {return 0;}
-int NoHostFs::Access(std::string name) {return 0;}
-Node* NoHostFs::ReadDir(int fd) {return 0;}
-int NoHostFs::DeleteFile(std::string name) {return 0;}
-int NoHostFs::DeleteDir(std::string name) {return 0;}
-int NoHostFs::CreateDir(std::string name) {return 0;}
-bool NoHostFs::DirExists(std::string name) {return 0;}
-int NoHostFs::GetFileSize(std::string name) {return 0;}
-int NoHostFs::GetFileModificationTime(std::string name) {return 0;}
-bool NoHostFs::Link(std::string src, std::string target) {return 0;}
-bool NoHostFs::IsEof(int fd) {return 0;}
-int NoHostFs::Lseek(int fd, uint64_t n) {return 0;}
+
+int NoHostFs::Lock(std::string name, bool lock){
+	return global_file_tree->Lock(name, lock);
+}
+
+int NoHostFs::Close(int fd){
+	if(fd < 0){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
+	OpenFileEntry* entry = open_file_table->at(fd);
+	if(entry == NULL){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
+	delete entry;
+	open_file_table->at(fd) = NULL;
+
+	return true;
+}
+int NoHostFs::Access(std::string name){
+	Node* node = global_file_tree->GetNode(name);
+	if(node == NULL){
+		errno = ENOENT; // No such file or directory
+		return -1;
+	}
+	return 1;
+}
+int NoHostFs::Rename(std::string old_name, std::string name){
+	Node* node = global_file_tree->GetNode(old_name);
+	if(node == NULL){
+		errno = ENOENT; // No such file or directory
+		return -1;
+	}
+	std::vector<std::string> path_list = split(name, '/');
+	node->name = path_list.back();
+
+	return 0;
+}
+Node* NoHostFs::ReadDir(int fd){
+	if(fd < 0){
+		errno = EBADF; // Bad file number
+		return NULL;
+	}
+	OpenFileEntry* entry = open_file_table->at(fd);
+	if(entry == NULL){
+		errno = EBADF; // Bad file number
+		return NULL;
+	}
+	uint64_t loop_count = 0;
+	if(entry->node->isfile){
+		errno = ENOTDIR; // not a directory
+		return NULL;
+	}
+	std::list<Node*>::iterator iter = entry->node->children->begin();
+
+	while(loop_count < entry->r_offset){
+		iter++;
+		loop_count++;
+	}
+	entry->r_offset++;
+	if(iter == entry->node->children->end())
+		return NULL;
+
+	return *iter;
+}
+int NoHostFs::DeleteFile(std::string name){
+	return global_file_tree->DeleteFile(name);
+}
+int NoHostFs::DeleteDir(std::string name){
+	return global_file_tree->DeleteDir(name);
+}
+int NoHostFs::CreateDir(std::string name){
+	if(global_file_tree->CreateDir(name) == NULL)
+		return -1;
+	return 1;
+}
+int NoHostFs::CreateFile(std::string name){
+	if(global_file_tree->CreateFile(name) == NULL)
+		return -1;
+	return 1;
+}
+bool NoHostFs::DirExists(std::string name){
+	if(Access(name) == -1) return false;
+	return true;
+}
+int NoHostFs::GetFileSize(std::string name){
+	Node* node = global_file_tree->GetNode(name);
+	if(node == NULL){
+		errno = ENOENT; // No such file or directory
+		return -1;
+	}
+	return node->GetSize();
+}
+uint64_t NoHostFs::GetFileModificationTime(std::string name){
+	Node* node = global_file_tree->GetNode(name);
+	if(node == NULL){
+		errno = ENOENT; // No such file or directory
+		return -1;
+	}
+	return node->last_modified_time;
+}
+bool NoHostFs::Link(std::string src, std::string target){
+	Node* node = global_file_tree->Link(src, target);
+	if(node == NULL) return -1;
+	return true;
+}
+bool NoHostFs::IsEof(int fd){
+	if(fd < 0){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
+	OpenFileEntry* entry = open_file_table->at(fd);
+	if(entry == NULL){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
+	if(entry->r_offset == entry->node->GetSize())
+		return true;
+	else
+		return false;
+}
+int NoHostFs::Lseek(int fd, uint64_t n){
+	if(fd < 0){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
+	OpenFileEntry* entry = open_file_table->at(fd);
+	if(entry == NULL){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
+	entry->r_offset = n;
+	entry->w_offset = n;
+	return entry->r_offset;
+}
 
 
 
@@ -22,21 +145,29 @@ int NoHostFs::Open(std::string name, char type){
 	Node* ret = NULL;
 	uint64_t start_address;
 
+	ret = global_file_tree->GetNode(name);
+
 	switch(type){
 	case 'd' :
-		break;
 	case 'r' :
-		ret = global_file_tree->GetNode(name);
+		if(ret == NULL){
+			errno = ENOENT; // No such file or directory
+			return -1;
+		}
 		break;
 	case 'w' :
-		global_file_tree->DeleteFile(name);
-		global_file_tree->CreateFile(name);
-		ret = global_file_tree->GetNode(name);
+		if(ret != NULL)
+			global_file_tree->DeleteFile(name);
+
+		ret = global_file_tree->CreateFile(name);
 		start_address = GetFreeBlockAddress();
 		ret->file_info->push_back(new FileSegInfo(start_address, 0));
 		break;
 	default:
 		std::cout << "r:readonly, w:read/write, you must choose one of them\n";
+		errno = EINVAL; // invalid argument
+		return -1;
+
 	}
 	open_file_table->push_back(new OpenFileEntry(ret, 0, 0));
 
@@ -44,9 +175,17 @@ int NoHostFs::Open(std::string name, char type){
 }
 
 int NoHostFs::Write(int fd, const char* buf, uint64_t size){
+	if(fd < 0){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
 	OpenFileEntry* entry = open_file_table->at(fd);
 	FileSegInfo* finfo = NULL;
 	int wsize = 0;
+	if(entry == NULL){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
 
 	uint64_t start_page = entry->w_offset / page_size;
 	uint64_t last_page = (entry->w_offset + size - 1) / page_size;
@@ -74,11 +213,20 @@ int NoHostFs::Write(int fd, const char* buf, uint64_t size){
 		entry->w_offset += wsize;
 	}
 
+	entry->node->last_modified_time = GetCurrentTime();
 	return wsize;
 }
 
 int NoHostFs::Write(int fd, char* buf, uint64_t size){
+	if(fd < 0){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
 	OpenFileEntry* entry = open_file_table->at(fd);
+	if(entry == NULL){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
 	FileSegInfo* finfo = NULL;
 	char* curbuf = buf;
 	int wsize = 0;
@@ -109,10 +257,15 @@ int NoHostFs::Write(int fd, char* buf, uint64_t size){
 		entry->w_offset += wsize;
 	}
 
+	entry->node->last_modified_time = GetCurrentTime();
 	return wsize;
 }
 
 int NoHostFs::SequentialRead(int fd, char* buf, uint64_t size){
+	if(fd < 0){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
 	std::string data = "";
 	char* tmp = new char[page_size];
 	int left = size;
@@ -120,19 +273,28 @@ int NoHostFs::SequentialRead(int fd, char* buf, uint64_t size){
 	int sum = 0;
 	while (left != 0) {
 		done = ReadHelper(fd, tmp, left);
-		if (done < 0) return false;
+		if (done < 0) return -1;
 		if(tmp != NULL) data += tmp;
 		left -= done;
 		sum += done;
 	}
+
 	strcpy(buf, data.data());
 	return sum;
 }
 
 int NoHostFs::ReadHelper(int fd, char* buf, uint64_t size){
+	if(fd < 0){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
 	OpenFileEntry* entry = open_file_table->at(fd);
 	FileSegInfo* finfo = NULL;
 	int rsize = 0;
+	if(entry == NULL){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
 
 	uint64_t start_page = entry->r_offset / page_size;
 	uint64_t last_page = (entry->r_offset + size - 1) / page_size;
@@ -161,9 +323,17 @@ int NoHostFs::ReadHelper(int fd, char* buf, uint64_t size){
 }
 
 int NoHostFs::Read(int fd, char* buf, uint64_t size){
+	if(fd < 0){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
 	OpenFileEntry* entry = open_file_table->at(fd);
 	FileSegInfo* finfo = NULL;
 	int rsize = 0;
+	if(entry == NULL){
+		errno = EBADF; // Bad file number
+		return -1;
+	}
 
 	uint64_t start_page = entry->r_offset / page_size;
 	uint64_t last_page = (entry->r_offset + size - 1) / page_size;
@@ -195,13 +365,13 @@ int NoHostFs::Read(int fd, char* buf, uint64_t size){
 uint64_t NoHostFs::GetFreeBlockAddress(){
 	uint64_t i;
 
-	for(i = 0; i < free_page_bitmap->size(); i++)
-		if(free_page_bitmap->at(i) == 0) break;
+	for(i = 0; i < global_file_tree->free_page_bitmap->size(); i++)
+		if(global_file_tree->free_page_bitmap->at(i) == 0) break;
 
-	if(i == free_page_bitmap->size())
-		free_page_bitmap->push_back(0);
+	if(i == global_file_tree->free_page_bitmap->size())
+		global_file_tree->free_page_bitmap->push_back(0);
 
-	free_page_bitmap->at(i) = 1;
+	global_file_tree->free_page_bitmap->at(i) = 1;
 	return i*page_size;
 }
